Move TUI activity formatting to adapter
